+++ Thermo/ThermoStaticData.cs
@@ -68,7 +68,10 @@ namespace IO.Thermo
 
             IThermoScan[] scans = new IThermoScan[pnLastSpectrum - pnFirstSpectrum + 1];
             for (int nScanNumber = pnFirstSpectrum; nScanNumber <= pnLastSpectrum; nScanNumber++)
-                scans[nScanNumber - pnFirstSpectrum] = GetMsDataOneBasedScanFromThermoFile(nScanNumber, theConnection, p, topNpeaks, minRatio, trimMs1Peaks, trimMsMsPeaks);
+            {
+                FilteringParams ThermoParams = new FilteringParams(topNpeaks,minRatio);
+                scans[nScanNumber - pnFirstSpectrum] = GetMsDataOneBasedScanFromThermoFile(nScanNumber, theConnection, p,ThermoParams, trimMs1Peaks, trimMsMsPeaks);
+            }
 
             theConnection.Close();
